Add form name param to graphql submission query
