+++ forms-flow-web/src/components/Dashboard/Dashboard.js
@@ -4,7 +4,6 @@ import ApplicationCounter from "./ApplicationCounter";
 import { useDispatch, useSelector } from "react-redux";
 import { Route, Redirect } from "react-router";
 import StatusChart from "./StatusChart";
-import StatusChart_POC from "./StatusChart-POC";
 import Modal from "react-bootstrap/Modal";
 import {
   fetchMetricsSubmissionCount,
@@ -359,23 +358,12 @@ const Dashboard = React.memo(() => {
                   </div>
                 </Modal.Header>
                 <Modal.Body>
-                  <div>
-                    <p>Redash</p>
                     <StatusChart
                       submissionsStatusList={submissionsStatusList}
                       submissionData={showSubmissionData}
                       getStatusDetails={getStatusDetails}
                       submissionStatusCountLoader={submissionStatusCountLoader}
                           />
-                    <hr></hr>
-                    <p>Charts.js</p>
-                    <StatusChart_POC
-                      submissionsStatusList={submissionsStatusList}
-                      submissionData={showSubmissionData}
-                      getStatusDetails={getStatusDetails}
-                      submissionStatusCountLoader={submissionStatusCountLoader}
-                          />
-                  </div>
                       </Modal.Body>
                     </Modal>
                   )}
