+++ forms-flow-web/src/components/Dashboard/StatusChart.js
@@ -1,58 +1,120 @@
-import React, { useMemo } from "react";
+import React, { useState } from "react";
+import { FormControl } from "react-bootstrap";
 import { useTranslation } from "react-i18next";
 import LoadingOverlay from "react-loading-overlay-ts";
 
-import { Legend, PieChart, Pie, Cell } from "recharts";
+import {
+  ArcElement,
+  BarElement,
+  CategoryScale,
+  Chart as ChartJS,
+  Filler,
+  Legend,
+  LinearScale,
+  LineElement,
+  PointElement,
+  RadialLinearScale,
+  Tooltip
+} from 'chart.js';
+import { Bar, Doughnut, Pie, PolarArea, Radar } from 'react-chartjs-2';
 
-const COLORS = [
-  "#0088FE",
-  "#00C49F",
-  "#FFBB28",
-  "#FF8042",
-  "#a05195",
-  "#d45087",
-  "#f95d6a",
-  "#ff7c43",
+ChartJS.register(
+  ArcElement,
+  BarElement,
+  CategoryScale,
+  Filler,
+  Legend,
+  LinearScale,
+  LineElement,
+  PointElement,
+  RadialLinearScale,
+  Tooltip
+);
+
+const chartTypes = [
+  { value: 'pie', label: 'Pie Chart' },
+  { value: 'v-bar', label: 'Vertical Bar Chart' },
+  { value: 'doughnut', label: 'Doughnut Chart' },
+  { value: 'polar-area', label: 'Polar Area Chart' },
+  { value: 'radar', label: 'Radar Chart' },
 ];
 
-const testData = [
-  {
-    "statusName": "Completed",
-    "count": 10
-  },
-  {
-    "statusName": "In Progress",
-    "count": 4
-  },
-  {
-    "statusName": "Inactive",
-    "count": 3
-  },
-  {
-    "statusName": "Other",
-    "count": 1
-  }
-];
+const testData = {
+  datasets: [
+    {
+      label: 'Submissions Dataset',
+      data: [{"Completed": 10}, {"In Progress": 4}, {"Inactive": 3}, {"Other": 1}],
+      backgroundColor: [
+        "#0088FE",
+        "#00C49F",
+        "#FFBB28",
+        "#FF8042",
+        "#a05195",
+        "#d45087",
+        "#f95d6a",
+        "#ff7c43"
+      ]
+    }
+  ]
+};
 
-// label={renderCustomizedLabel}
+
 const ChartForm = React.memo((props) => {
   const { submissionsStatusList, submissionData, submissionStatusCountLoader } = props;
-  const {formVersions, formName, parentFormId} = submissionData;
+  const { formName } = submissionData;
 
-  const sortedVersions = useMemo(()=>
-  (formVersions?.sort((version1, version2)=>
-  version1.version > version2.version ? 1 : -1)),[formVersions]);
+  const [selectedChartValue, setSelectedChartValue] = useState('pie');
 
-  const version = formVersions?.length;
   const { t } = useTranslation();
-  let pieData = submissionsStatusList || [];
-  pieData = testData;
+  const chartData = submissionsStatusList || [];
+  console.log(chartData);
 
-  const handlePieData = (value) => {
-    const isParentId = value === "all";
-    const id = isParentId ? parentFormId : value;
-    const option = {parentId : isParentId};
-    props.getStatusDetails(id,option);
+  const chartOptions = {
+    plugins: {
+      legend: {
+        position: 'bottom'
+      }
+    }
+  };
+
+  const renderChart = () => {
+    switch (selectedChartValue) {
+      case 'pie':
+        return (
+          <Pie
+            data={testData}
+            options={chartOptions}
+          />
+        );
+      case 'v-bar':
+        return (
+          <Bar
+            data={testData}
+            options={chartOptions}
+          />
+        );
+      case 'doughnut':
+        return (
+          <Doughnut
+            data={testData}
+            options={chartOptions}
+          />
+        );
+      case 'polar-area':
+        return (
+          <PolarArea
+            data={testData}
+            options={chartOptions}
+          />
+        );
+      case 'radar':
+        return (
+          <Radar
+            data={testData}
+            options={chartOptions}
+          />
+        );
+    }
   };
 
 
@@ -72,90 +134,33 @@ const ChartForm = React.memo((props) => {
                   {formName}
                 </h2>
               </div>
-
-          <p>
-            <span className="text-primary" >{t("Latest Version")} :</span>{" "}
-            {`v${version}`}
-          </p>
-          </div>
-          {sortedVersions.length > 1 && (
-  <div className="d-flex align-items-center">
-    <p className="text-primary mb-0 me-2" style={{ whiteSpace: "nowrap" }}>
-      {t("Select form version")}:
-    </p>
-    <select
-      className="form-select"
-      aria-label="Default select example"
-      onChange={(e) => {
-        handlePieData(e.target.value);
-      }}
-    >
-      {sortedVersions.map((option) => (
-        <option key={option.formId} value={option.formId}>
-          v{option.version}
-        </option>
-      ))}
-      <option selected value={"all"}>
-        {t("All")}
-      </option>
-    </select>
-  </div>
-)}
-
-
+            </div>
           </div>
           <LoadingOverlay
-        active={submissionStatusCountLoader}
-        spinner
-        text={t("Loading...")}
-      >
-         <div className="white-box status-container flex-row d-md-flex flex-wrap align-items-center justify-content-around">
- 
-    <div className="col-md-6">
-      <PieChart width={400} height={400}>
-        <Pie
-          paddingAngle={1}
-          minAngle={1}
-          data={pieData}
-          labelLine={false}
-          outerRadius={90}
-          fill="#8884d8"
-          dataKey="count"
-          nameKey="statusName"
-          label
-        >
-          <Legend />
-
-          {pieData.map((entry) => (
-  <Cell key={entry.statusName} fill={COLORS[pieData.indexOf(entry) % COLORS.length]} />
-))}
-        </Pie>
-      </PieChart>
-    </div>
-    {
-              pieData.length ? (
-                <div className="d-flex border flex-wrap rounded p-4   ">
-              {pieData.map((entry, index) => (
-                <div className=" d-flex align-items-center m-3" key={entry.statusName}>
-                  <span
-                    className="rounded-circle shadow  me-2"
-                    style={{
-                      backgroundColor: COLORS[index % COLORS.length],
-                      width: "25px",
-                      height: "25px",
-                    }}
-                  ></span>
-                  <div className="legent-text">{entry.statusName}</div>
+            active={submissionStatusCountLoader}
+            spinner
+            text={t("Loading...")}
+          >
+            <div className="white-box status-container flex-row d-md-flex flex-wrap align-items-center justify-content-around">
+              <div className="col-md-6">
+                {renderChart()}
+              </div>
+              <div className="col-md-3 offset-md-1">
+                <div className="input-group">
+                  <FormControl
+                    as="select"
+                    onChange={(e) => setSelectedChartValue(e.target.value)}
+                    className="form-select p-1"
+                    title={t("Choose any")}
+                    aria-label="Select chart type"
+                  >
+                    {chartTypes.map((option, index) => (
+                      <option key={index} value={option.value}>{option.label}</option>
+                    ))}
+                  </FormControl>
                 </div>
-              ))}
+              </div>
             </div>
-              )  : (
-    <div className="d-flex justify-content-center align-items-center w-100" style={{ minHeight: "200px" }}>
-      <span className="text-center">{t("No submissions")}</span>
-    </div>
-  )}
-</div>
-
           </LoadingOverlay>
         </div>
       </div>
